+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.11.0
+## 1.11.0 (November 2, 2019)
 * Added support for HEIF images.
 * Added [`--hcompression`](https://sharp.pixelplumbing.com/en/stable/api-output/#heif) option.
 * Added [`--premultiplied`](https://sharp.pixelplumbing.com/en/latest/api-composite/#composite) option.